Aggregate some routes
