+++ app/chatpal-search/client/index.js
@@ -4,5 +4,4 @@ import './template/suggestion.html';
 import './template/admin';
 import './template/result';
 import './template/suggestion';
-import './route';
 import './style.css';
