+++ app/message-snippet/client/index.js
@@ -1,7 +1,6 @@
 import './actionButton';
 import './messageType';
 import './snippetMessage';
-import './router';
 import './page/snippetPage.html';
 import './page/snippetPage';
 import './tabBar/tabBar';
