+++ client/views/admin/routes.js
@@ -1,5 +1,7 @@
 import { Meteor } from 'meteor/meteor';
+import { lazy } from 'react';
 
+import { appLayout } from '../../lib/appLayout';
 import { createRouteGroup } from '../../lib/createRouteGroup';
 
 export const registerAdminRoute = createRouteGroup('admin', '/admin', () => import('./AdministrationRouter'));
@@ -128,4 +130,13 @@ Meteor.startup(() => {
 		name: 'admin',
 		lazyRouteComponent: () => import('./settings/SettingsRoute'),
 	});
+});
+
+const MainLayout = lazy(() => import('../root/MainLayout'));
+
+registerAdminRoute('/chatpal', {
+	name: 'chatpal-admin',
+	action() {
+		appLayout.render({ component: MainLayout, props: { center: 'ChatpalAdmin' } });
+	},
 });
