+++ ee/client/startup/index.ts
@@ -1,1 +1,2 @@
 import './engagementDashboard';
+import './routes';
