+++ ee/client/startup/routes.ts
@@ -1,9 +1,9 @@
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { lazy } from 'react';
 
-import { appLayout } from '../../../../client/lib/appLayout';
+import { appLayout } from '../../../client/lib/appLayout';
 
-const MainLayout = lazy(() => import('../../../../client/views/root/MainLayout'));
+const MainLayout = lazy(() => import('../../../client/views/root/MainLayout'));
 
 FlowRouter.route('/audit', {
 	name: 'audit-home',
